2020-07-29修改Viewport的宽高为int类型；更改删除缓存文件夹的方式；对pdf方法添加判断空指针问题

+++ src/main/java/com/ruiyun/jvppeteer/core/page/ElementHandle.java
@@ -162,8 +162,8 @@ public class ElementHandle extends JSHandle {
         Viewport viewport = this.page.viewport();
         if (viewport != null && (boundingBox.getWidth() > viewport.getWidth() || boundingBox.getHeight() > viewport.getHeight())) {
             Viewport newViewport = new Viewport();
-            newViewport.setWidth(Math.max(viewport.getWidth(), Math.ceil(boundingBox.getWidth())));
-            newViewport.setHeight(Math.max(viewport.getHeight(), Math.ceil(boundingBox.getHeight())));
+            newViewport.setWidth(Math.max(viewport.getWidth(), (int)Math.ceil(boundingBox.getWidth())));
+            newViewport.setHeight(Math.max(viewport.getHeight(), (int)Math.ceil(boundingBox.getHeight())));
 
             this.page.setViewport(newViewport);
             needsViewportReset = true;
